+++ nnetsauce/lazypredict/__init__.py
@@ -1,13 +1,12 @@
-from .lazyClassifier import LazyClassifier
-from .lazyRegressor import LazyRegressor
-from .lazydeepClassifier import LazyDeepClassifier
-from .lazydeepRegressor import LazyDeepRegressor
-from .lazydeepMTS import LazyDeepMTS
+from .lazydeepClassifier import LazyDeepClassifier, LazyClassifier
+from .lazydeepRegressor import LazyDeepRegressor, LazyRegressor
+from .lazydeepMTS import LazyDeepMTS, LazyMTS
 
 __all__ = [
     "LazyClassifier",
     "LazyRegressor",
     "LazyDeepClassifier",
     "LazyDeepRegressor",
+    "LazyMTS",
     "LazyDeepMTS",
 ]